+++ models/api/imotion/IMotionUpdateSupporter.php
@@ -4,11 +4,14 @@ declare(strict_types=1);
 
 namespace app\models\api\imotion;
 
+use app\models\db\ISupporter;
+
 class IMotionUpdateSupporter
 {
     public function __construct(
         public string $name,
         public ?int $id = null,
+        public ?SupporterType $personType = null,
         public ?string $organization = null,
         public ?string $gender = null,
     ) {
@@ -32,9 +35,12 @@ class IMotionUpdateSupporter
             $organization = isset($post['supporters']['organization'][$i])
                 ? trim((string)$post['supporters']['organization'][$i])
                 : null;
+            $personType = (intval($post['supporters']['personType'][$i] ?? ISupporter::PERSON_NATURAL) === ISupporter::PERSON_ORGANIZATION ?
+                SupporterType::ORGANIZATION : SupporterType::PERSON);
             $supporters[] = new self(
                 id: isset($post['supporters']['id']) && $post['supporters']['id'][$i] > 0 ? intval($post['supporters']['id'][$i]) : null,
                 name: $name,
+                personType: $personType,
                 organization: $organization !== '' ? $organization : null,
                 gender: $post['supporters']['gender'][$i] ?? null,
             );